requirements.txt
beautifulsoup4
requests
bs4
html5lib
urllib3
python-telegram-bot
